+++ src-executables/Main-trident.hs
@@ -31,7 +31,8 @@ import           Poseidon.CLI.Validate                   (ValidateOptions (..),
 import           Poseidon.Janno                          (jannoHeaderString)
 import           Poseidon.PoseidonVersion                (showPoseidonVersion,
                                                           validPoseidonVersions)
-import           Poseidon.Utils                          (LogMode (..),
+import           Poseidon.Utils                          (ErrorLength (..),
+                                                          LogMode (..),
                                                           PlinkPopNameMode (..),
                                                           PoseidonException (..),
                                                           PoseidonIO, TestMode,
@@ -77,17 +78,12 @@ main = do
     hPutStrLn stderr renderVersion
     hPutStrLn stderr ""
     (Options logMode testMode errLength plinkMode subcommand) <- OP.customExecParser (OP.prefs OP.showHelpOnEmpty) optParserInfo
-    catch (usePoseidonLogger logMode testMode plinkMode $ runCmd subcommand) (handler logMode testMode errLength plinkMode)
+    catch (usePoseidonLogger logMode testMode plinkMode errLength $ runCmd subcommand) (handler logMode testMode plinkMode errLength)
     where
-        handler :: LogMode -> TestMode -> ErrorLength -> PlinkPopNameMode -> PoseidonException -> IO ()
-        handler l t len pm e = do
-            usePoseidonLogger l t pm $ logError $ truncateErr len $ renderPoseidonException e
+        handler :: LogMode -> TestMode -> PlinkPopNameMode -> ErrorLength -> PoseidonException -> IO ()
+        handler l t pm len e = do
+            usePoseidonLogger l t pm len $ logError $ renderPoseidonException e
             exitFailure
-        truncateErr :: ErrorLength -> String -> String
-        truncateErr CharInf         s = s
-        truncateErr (CharCount len) s
-            | length s > len          = take len s ++ "... (see more with --errLength)"
-            | otherwise               = s
 
 runCmd :: Subcommand -> PoseidonIO ()
 runCmd o = case o of