+++ src/Poseidon/CLI/OptparseApplicativeParsers.hs
@@ -23,7 +23,10 @@ import           Poseidon.GenotypeData      (GenoDataSource (..),
                                              SNPSetSpec (..))
 import           Poseidon.ServerClient      (AddJannoColSpec (..),
                                              ArchiveEndpoint (..))
-import           Poseidon.Utils             (LogMode (..), TestMode (..))
+import           Poseidon.Utils             (ErrorLength (..), LogMode (..),
+                                             TestMode (..),
+                                             renderPoseidonException,
+                                             showParsecErr)
 import           Poseidon.Version           (VersionComponent (..),
                                              parseVersion)
 
@@ -84,7 +87,7 @@ parseMaybePoseidonVersion = OP.option (Just <$> OP.eitherReader readPoseidonVers
     where
         readPoseidonVersionString :: String -> Either String Version
         readPoseidonVersionString s = case P.runParser parseVersion () "" s of
-            Left p  -> Left (show p)
+            Left p  -> Left (showParsecErr p)
             Right x -> Right x
 
 parseDebugMode :: OP.Parser LogMode
@@ -129,13 +132,11 @@ parseTestMode = OP.option (OP.eitherReader readTestMode) (
             "Production" -> Right Production
             _            -> Left "must be Testing or Production"
 
-data ErrorLength = CharInf | CharCount Int deriving Show
-
 parseErrorLength :: OP.Parser ErrorLength
 parseErrorLength = OP.option (OP.eitherReader readErrorLengthString) (
     OP.long "errLength" <>
     OP.metavar "INT" <>
-    OP.help "After how many characters should a potential error \
+    OP.help "After how many characters should a potential genotype data parsing error \
             \message be truncated. \"Inf\" for no truncation." <>
     OP.value (CharCount 1500) <>
     OP.showDefault
@@ -231,7 +232,7 @@ parseContributors = OP.option (OP.eitherReader readContributorString) (
 
 readContributorString :: String -> Either String [ContributorSpec]
 readContributorString s = case P.runParser contributorSpecParser () "" s of
-    Left p  -> Left (show p)
+    Left p  -> Left (showParsecErr p)
     Right x -> Right x
 
 
@@ -304,7 +305,7 @@ parseForgeEntitiesDirect = OP.option (OP.eitherReader readSignedEntities) (
         \source package, they can be specified with the special syntax \"<package:group:individual>\".")
   where
     readSignedEntities s = case readEntitiesFromString s of
-        Left e  -> Left (show e)
+        Left e  -> Left $ renderPoseidonException e
         Right e -> Right e
 
 parseFetchEntitiesDirect :: OP.Parser EntitiesList
@@ -319,7 +320,7 @@ parseFetchEntitiesDirect = OP.option (OP.eitherReader readEntities) (
         \specified, then packages which include these groups or individuals are included in the download.")
   where
     readEntities s = case readEntitiesFromString s of
-        Left e  -> Left (show e)
+        Left e  -> Left $ renderPoseidonException e
         Right e -> Right e
 
 parseForgeEntitiesFromFile :: OP.Parser FilePath