+++ src/Poseidon/EntityTypes.hs
@@ -20,7 +20,7 @@ module Poseidon.EntityTypes (
     resolveEntityIndices, reportDuplicateIndividuals) where
 
 import           Poseidon.Utils         (PoseidonException (..), PoseidonIO,
-                                         logError)
+                                         logError, showParsecErr)
 import           Poseidon.Version       (parseVersion)
 
 import           Control.Applicative    ((<|>))
@@ -230,7 +230,7 @@ instance ToJSON   SignedEntity   where toJSON e = String (pack $ show e)
 
 aesonParseEntitySpec :: (EntitySpec e) => Text -> Parser e
 aesonParseEntitySpec t = case P.runParser entitySpecParser () "" (unpack t) of
-    Left err -> fail (show err)
+    Left err -> fail $ showParsecErr err
     Right p' -> return p'
 
 -- | a minimal datatype representing an individual in a collection of packages
@@ -322,7 +322,7 @@ checkIfAllEntitiesExist entities indInfoCollection = do
 -- parsing code to read entities from files
 readEntitiesFromString :: (EntitySpec a) => String -> Either PoseidonException [a]
 readEntitiesFromString s = case P.runParser (entitiesListP <* P.eof) () "" s of
-    Left p  -> Left $ PoseidonPoseidonEntityParsingException (show p)
+    Left p  -> Left $ PoseidonPoseidonEntityParsingException p
     Right x -> Right x
 
 readEntityInputs :: (MonadIO m, EntitySpec a, Eq a) => [EntityInput a] -> m [a] -- An empty list means that entities are wanted.
@@ -363,5 +363,5 @@ readEntitiesFromFile :: (EntitySpec a) => FilePath -> IO [a]
 readEntitiesFromFile entitiesFile = do
     eitherParseResult <- P.parseFromFile (entitiesListMultilineP <* P.eof) entitiesFile
     case eitherParseResult of
-        Left err -> throwIO (PoseidonPoseidonEntityParsingException (show err))
-        Right r  -> return r
+        Left e  -> throwIO (PoseidonPoseidonEntityParsingException e)
+        Right r -> return r