+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+- V 1.5.4.0:
+    - Better error messages for parsec operations on command line input.
+    - Better error messages for YML file parsing.
+    - Refactored `--errLength` to exclusively cover exceptions of type `PoseidonGenotypeExceptionForward` emerging during genotype data parsing.
 - V 1.5.3.0:
     - Introduced a new output option `forge --preservePyml` to preserve certain features of the input package, in case there is only one.
     - Refactored the out mode selection in `forge` to replace an increasingly complex boolean logic with a clear separation of modes based on the new type `ForgeOutMode`.